data_scripts/requirements.txt
requests
beautifulsoup4
selenium
pandas
psycopg2-binary
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
